SWBCore: make the module build on Windows
`KF_KNOWN_FOLDER` is a strongly typed enumerator and does not import
into Swift as an integral type. Provide a local overload to allow
writing the code idiomatically.

`SUCCEEDED` is a "complex" macro that is beyond the ability of the
ClangImporter to import into Swift. Provide a trivial implementation of
the CPP macro in Swift to allow writing code idiomatically.