rdar://143411533 (Object files and module wrap output end up in the same location, stomping on each other) (#31)
On non macho-o platforms, swift-driver will create an additional module-wrap task if debugging is enabled which has an object file as output that will be named like the module and placed next to it. Since Swift Build will emit the Swift module next to the object files of a target, there can be a naming conflict, leading to us only ending up with the module-wrap content and the actual object file being stomped on.

This introduces a new `PER_ARCH_MODULE_FILE_DIR` build setting that's the same as `PER_ARCH_OBJECT_FILE_DIR` by default that allows us to customize the output directory.