Tests: repair the build on Windows
The adoption of Swift 6 triggered a failure on Windows. `HANDLE` is a
typealias for a pointer. This value is sendable. As this is a typealias
to `UnsafePointer<VOID__>` we require that the sendable conformance is
in the `Swift` module. However, that is not possible. Provide a
retroactive conformance to make the capture concurrency safe.