Fix PIFLoadingUnitTests.loadingStandardTarget() and clean up other tests in that suite.
loadingStandardTarget() had several checks which were failing, and had been broken for a long time but had been masked by other behavior which was recently fixed.

Many other tests were written early in the project's history, and could be updated to use newer paradigms. In particular, many unnecessary optionals and force-unwraps have been removed or converted.

One of the principles applied here is that tests should strive to perform as much as their testing as is reasonable, and not simply stop on first failure. Even lines which fail an #expect() can often move on to the section block of the test rather than skipping everything afterwards.