Fetching built target descriptions can take much longer than necessary <rdar://143802224> (#96)
Higher layer APIs properly request built target descriptions as a batch, but down here in the build system it still does a serialized fetch for each target ID that kicks off a build description request and rewalking all the targets to find the right one. When the build system is under contention (when indexing, for instance) this can cause target description requests to take much longer than necessary.

This change plumbs the batch target fetching API all the way down to the `BuildDescriptionManager`. 20x speedups observed in average cases!