Revert "Fix accidentally-introduced warnings and regression in test behavior in PIFLoadingTests"
The original code was correct and these changes overlooked some context as to why the original code was the way it was (e.g., the warnings are due to a compiler bug, #require will already emit an issue).

But we can work around most of the warnings by casting to an optional, which is low-profile and allows us to retain the same approach to these checks, so that has been added where possible.

This reverts commit a6a6bb748e130ef75b39d5d2bc039108d82497b8.