Fix some issues with extended attributes
On Linux, the name of an extended attribute "must be a null-terminated string prefixed by a namespace identifier and a dot character", and only the "user" namespace is available for unrestricted use.

This changes our use of extended attributes to prefer a Linux-compatible name where possible, and updates the unit tests to account for this restriction and enables that test on Linux.

This also clarifies the support situation for extended attributes on Windows and OpenBSD. The logical equivalent to xattrs on Windows is perhaps ADS; add a comment to that effect and a reference to the Foundation issue tracking potentially adding the support at that level. Compile out xattrs support on OpenBSD as they don't exist there at all.