Infer active Xcode for testing
Historically, we have been relying on `XCODE_DEVELOPER_DIR_PATH` being set to `DEVELOPER_DIR` in our schemes to find the active Xcode, but with auto-generated package schemes, this doesn't really work anymore. Try to infer the path from `PATH` instead which should be set when running in the context of Xcode.