Always add linkage dependencies to host tool targets for the index arena
Settings are propagated to dependent targets through the linkage graph
rather than the target graph (unless
`UseTargetDependenciesForImpartedBuildSettings` is set). While the
target graph was updated to allow host tool dependencies even when
configuring for a different platform in the index arena, the linkage
graph was not. This meant that we'd miss imparting settings in this case
(eg. include paths up from dependencies) and thus fail to build the host
tool in cases this is required. As the host tool is built without the
preparation settings (since it needs to actually build the executable),
this would then continually run for every preparation request.

The fix here is to add linkage dependencies for host tools as well and
to unify the two "is compatible for index arena" checks to avoid this in
the future. That exposes another issue, however, which is that doing so
would then result in having multiple configured targets all with the
same (host) platform. Avoid that by forcing the parameters to that of
the host platform for the index arena.

Resolves rdar://140882270.