Remove client-side API for macro evaluation scopes.
This primarily involves removing the messages to create and discard the scope, and client-side methods for doing so. Clients no longer hold on to a scope directly but rather request the service perform evaluations using a set of build parameters, which eliminates the need to do bookkeeping for these objects.

(There is still service-side caching of registered Settings objects in the Session, as it it used in PlanningOperation.)

rdar://142511013