Ensure we do not expect additional linker args to be produced in targets with Swift-produced object files (#378)
If Swift produced an object in the linked libraries of a target, we allow it to generate additional linker
options, even if it did not run to compile any source files. However, we should not attempt to inject top-level
AST paths for debugging since no actual compile took place, and the paths won't exist at execution time. Use the
object file type when injecting Swift tool usage in these targets so that we don't take the unintended code path.

rdar://144729550