Remove support for compiling with bitcode.
This removes support for compiling files with bitcode from SWBuild, as this functionality is no longer supported. This removes a lot of functionally dead code.

Support for stripping bitcode from copied files remains, as there are extant libraries and XCFrameworks which contain bitcode, which must be stripped. Bitcode was only supported for a subset of platforms, to the assignment of STRIP_BITCODE_FROM_COPIED_FILES in Settings based on platform remains, so time isn't spent stripping nonexistent bitcode from other platforms.

Some interfaces which Xcode uses continue to refer to bitcode for compatibility, but effectively do nothing.

Note that the references to bitcode in the OpenCL support are a concept internal to that tool, and independent of the support being removed here.

rdar://148287993