Impose `PATH` onto linker driver
Currently, we use Swift Build's search paths for determining which linker will be used and for version discovery etc, but we are not making any effort to actually enforce that the linker tools found using search paths are the ones used in practice. In the majority of cases, the used toolchain comes with linker tools and linker drivers will use those by default, so the discrepancy doesn't matter, but for toolchains where this isn't the case, what Swift Build thinks the linker will be and what it actually is will diverge.

This changes the environment in of linker tools to fix that discrepancy.