Check for the index workspace description rather than index arena
There's various paths that should be specific to the index workspace
description rather than if the index arena is enabled. Specifically the
recent change to replace parameters for host tools
(https://github.com/swiftlang/swift-build/pull/345) broke the *package*
description as it uses the passed in run destination rather than
configuring for all platforms. This replacement would then create two
configured targets for macOS for the same target (one for the given run
destination and one for the macOS platform that the index uses).

This also changes the *target* description to only build for the run
destination, which makes sense given its entire purpose is to be a
replacement for the workspace description if it hasn't built yet (ie. it
should be as fast as possible).

Resolves rdar://149896775.