Remove confusing SWBDispatchQueue.async overload
This overload was only meant to be used in one place, and its behavior was a little confusing, since it does not actually run the work on the queue. Over time, this overload got used in a few other places due to overload resolution picking it up, and turning formerly synchronous work asynchronous. Remove it to restore the intended behavior.