Fix typo in `isCompatible`
This just looks like a straight typo since `SpecializationParameters.toolchain` corresponds to the `TOOLCHAINS` build setting. I'm a bit surprised we have not seen this materialize in any bugs, but then again, customizing `TOOLCHAINS` is somewhat rare and hitting this would require two targets with different customizations of `TOOLCHAINS` to be depending on the same specialized target for the same platform.