Handle multiple llvm-cas execs for validation
While we intentionally wanto to ignore minor differences in CASOptions
for validation, the llvm-cas executable path is significant because
there could be differences in format version between tools. If the
llvm-cas binaries end up using the same format version only one of them
will perform the real validation and the rest will skip.