[cas] Add ValidateCAS action to ensure data coherence (#460)
Use llvm-cas's new validate-if-needed action to ensure the correctness
of CAS data in the case of power failure or similar situations. This
action is added to prepareForBuilding to ensure that it is run before
any other other CAS accesses. Note that validate-if-needed internally
avoids performing unnecessary work - in particular, it only validates
data once for every machine boot.

rdar://150295950