Add swift version file to record the Swift toolchain to use
When working with SwiftBuild, there is a particular version that
development is expected to use so that all of the compilers, standard
libraries, test frameworks, etc. are expected to be working. The
`.swift-version` file establishes a standard way to record this
information. Also, there is tooling available through `swiftly` to use
that information to automatically use the correct toolchain in any git
commit to set the developer's toolchain when they run and test the
package functions at desk.

Create the `.swift-version` file and set it to the current expected
toolchain version: 6.1.