[Explicit Module Builds] Add a setting to allow creating a reproducer for errors.
The idea is the same as Clang flag `-gen-reproducer=error` but for
explicit module builds. The name of the setting is a subject to change
depending on how it ends up being used.

Not all module-related issues are crashes, so it is useful to allow
creating and sharing reproducers for errors.

rdar://158780565