Generalize the Swift version regex
The "swiftlang" version was only being used in a single location for a
compatibility check that is no longer required. The clang version was
unusued (which is unsurprising since it's often not available from the
Swift `--version` output). Really we only need the version itself and a
unique tag to pass as additional signature data to Swift tasks.