Don't require objc in modulemaps generated for Swift compatibility headers
Let's remove the `requires objc` from the module map file generated for
the compatibility header. This condition triggered an error when
importing the compatibility header from a C source file.

This declaration is superfluous, the compatibility header is already
printed in a way where the Objective-C code is protected behind a
language check. C clients can safely import the current compatibility
header even if they may not see any content.

Let's lift this restriction. It isn't currently necessary and we're
adding C content to the compatibility header with the official support
for `@cdecl` that is independent of Objective-C.