Remove duplicate module names in diagnostic about missing dependencies
The Clang tracing information currently shows each header file separately,
so if there are multiple headers included from the same module, we get
duplicates. We are investigating having Clang trace the modules directly
and also investigating de-duplicating these diagnostics across targets, but
this is easy to fix in the meantime.

rdar://156160529