Disable modular builds when compiling for >= c++20 modes
The clang compiler will implicitly enable the modules feature when in
language mode c++20 or greater. This is for c++20 modules which strictly
differs from clang modules, but are conflated at times like when the
compiler evaluates `__has_feature(modules)`.
Since the sdks cannot cleanly build in this mode, disable it by default
but allow explicit overrides.

rdar://156935289