Add HEADER_DEPENDENCIES build setting with validation
This is similar to MODULE_DEPENDENCIES but for non-modular header file inclusion
dependencies. Our approach to this has changed a bit over the last few months,
and the trace output from Clang does not distinguish the modular vs. non-modular
includes. We will work on fixing that, but in the meantime, this continues the
placeholder approach used for MODULE_DEPENDENCIES where it assumes anything in a
framework has a module with the same name as the framework.

This also removes moduleDependenciesContext from ClangTaskPayload because the
code that needs that now gets it from ValidateDependenciesPayload.

rdar://150309197