Only check `isTargetSuitableForPlatformForIndex` for workspace description (#671)
* [test] NFC: Factor out `withHostToolsPackages`

Factor out the creation of the common package dependencies between
these two tests.

* [test] Parameterize two host tools index tests

Use Swift Testing parameterization to test different run destinations.

* [test] Set `SDKROOT: auto` at project level for `withHostToolsPackages`

This better matches what we actually generate for packages. Also
set all available platforms as supported.

* [test] Add same-package host tool dependency to `withHostToolsPackages`

This exposes the issue where `isTargetSuitableForPlatformForIndex`
returns `false` in the target and package build. Also change
`testHostToolsAndDependenciesAreBuiltDuringIndexingPreparationForPackage`
to use a dependency package and test both the target and package
build descriptions.

* Only check `isTargetSuitableForPlatformForIndex` for workspace description

For the package and target build description we can end up incorrectly
dropping build tool dependencies since not all clients pass in the
correct dependency information. It's not clear that we actually gain
much from doing this check for the target and package build
descriptions though, its primary purpose is to avoid configuring
unsupported targets in the workspace build description where we try
to configure for all available platforms. As such, switch to only
checking for the workspace build description. Note we don't encounter
this issue in the workspace case since there we override the build
parameters for host build tools to always build for the host platform.

rdar://152012769

* Remove dependency check from `isTargetSuitableForPlatformForIndex`

This gets applied pretty inconsistently, only 1 client actually
passes the correct dependency information. We ought to be able to rely
on the host platform being imposed for the workspace build description,
so it shouldn't be necessary.