Still emit per-task C diagnostics if errors are requested
We want to group diagnostics to avoid duplicates, but also avoid building the entire target if we already know about errors. So for C compiles specifically, we'll continue emitting diagnostics per-task if errors are enabled.

Additionally, this PR fixes up the signature data related to dependency diagnostics. The `ValidateDependencies` task needs to depend on the `MODULE_DEPENDENCIES` build setting, but the Swift planning task shouldn't anymore. For Clang tasks, this is now also dependent on whether errors are enabled or not.