Upgrade SWBUtil to Swift 6
This also required a bit of a rethink of the plugin manager now that actor isolation checking is enabled for types defined in SWBUtil. We introduce a new plugin manager type and cover protocol representing an immutable view of the plugin loading state. This is constructed from the mutable plugin manager just prior to core initialization, ensuring Core and everything underneath it have a completely immutable view of the plugin loading state, without having to worry about async/actors, making all of the extension points accessible from synchronous nonisolated code.