Add a request to find a most appropriate configured target for indexing purposes
It's prompted by the fact that LSP has to operate on targets instead
of configured targets at the moment because there is (currently) no
way (or enough information) on LSP side to select a canonical target
for a given document.

This change creates an API endpoint for `selectConfiguredTargetForIndex`.
It makes it possible to select a single configured target given a target GUID and
a run destination (+ other parameters).