Merge dumped dependency info
Since #874, we are dumping per-target dependency information, this adds a new task which merges these for a given build request. The task is only created if there's at least one target in the graph that will dump dependency information, so should be a no-op for all builds right now.

Additionally, this declares the dependency info file as an output of `ValidateDependencies` which I neglected to do in my earlier PR.