Add fallback search paths for Unix utilities
On macOS, some projects manipulate PATH and become unable to find chown due to the recent changes. To protect against this and similar future changes, add a set of default search paths for the typical Unix tool locations of /usr/{bin,sbin} and /{bin,sbin}, attempting to follow the usual relative orderings on each platform.