Swift SDK for WASM using the run destination
The client can provide a Swift SDK manifest path along with a
triple in the run destination to synthesize a Swift Build SDK.
This serves as a fallback for cases where Swift Build cannot
match an SDK for it.

Note that in this first implementation the platform is hard
coded to webassembly, and so only SDK's that align with WASM,
such as static linking of compiler runtime pieces, and some
Unix semantics will work.