Move `CODEOWNERS` to `.github/CODEOWNERS`
The Contributor Experience Workgroup is currently formalizing the Code Owner role and as part of that standardizes the definition of code owners in a .github/CODEOWNERS file across the swiftlang organization (https://github.com/swiftlang/swift-org-website/pull/1269). Move the existing `CODEOWNERS` file at the root of the repository into the `.github` folder to reflect this standard.